requirements.txt
altgraph==0.17.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
Faker==19.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
idna==3.4
oauthlib==3.2.2
pefile==2023.2.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.16
